Refactor: extract converter layer and clean up result classes
- Extract MarkdownConverter, WordConverter, LatexConverter from mixin.py
  into dedicated modules under common/result/converter/
- Extract format functions into format_funcs.py with clear naming
- Eliminate handle_funcs_dict build duplication via build_handle_funcs_dict()
- Remove inspect.signature hack; use explicit converter APIs
- Remove pass-through fields from MarkdownConverter.convert()
- Slim down result_v2.py and pp_doctranslation/result.py by delegating
  to converter layer
