+++ paddlex/inference/common/result/__init__.py
@@ -17,6 +17,7 @@ from .base_cv_result import BaseCVResult
 from .base_result import BaseResult
 from .base_ts_result import BaseTSResult
 from .base_video_result import BaseVideoResult
+from .converter import LatexConverter, MarkdownConverter, WordConverter
 from .mixin import (
     Base64Mixin,
     CSVMixin,
